cover input chain in codegen
