+++ ksp/src/main/kotlin/eu/vendeli/ksp/HelperUtils.kt
@@ -1,13 +1,22 @@
 package eu.vendeli.ksp
 
+import com.google.devtools.ksp.processing.Resolver
 import com.google.devtools.ksp.symbol.KSAnnotation
+import com.google.devtools.ksp.symbol.KSClassDeclaration
+import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import com.google.devtools.ksp.symbol.KSType
 import com.google.devtools.ksp.symbol.KSValueArgument
+import com.squareup.kotlinpoet.CodeBlock
+import com.squareup.kotlinpoet.FileSpec
+import com.squareup.kotlinpoet.KModifier
+import com.squareup.kotlinpoet.ParameterizedTypeName
+import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeVariableName
 import com.squareup.kotlinpoet.asTypeName
 import eu.vendeli.tgbot.TelegramBot
 import eu.vendeli.tgbot.types.User
 import eu.vendeli.tgbot.types.internal.CallbackQueryUpdate
+import eu.vendeli.tgbot.types.internal.ChLink
 import eu.vendeli.tgbot.types.internal.ChannelPostUpdate
 import eu.vendeli.tgbot.types.internal.ChatJoinRequestUpdate
 import eu.vendeli.tgbot.types.internal.ChatMemberUpdate
@@ -23,6 +32,10 @@ import eu.vendeli.tgbot.types.internal.PreCheckoutQueryUpdate
 import eu.vendeli.tgbot.types.internal.ProcessedUpdate
 import eu.vendeli.tgbot.types.internal.ShippingQueryUpdate
 import eu.vendeli.tgbot.types.internal.UpdateType
+import eu.vendeli.tgbot.types.internal.configuration.RateLimits
+import kotlin.reflect.KClass
+
+internal typealias FileBuilder = FileSpec.Builder
 
 internal val invocableType = TypeVariableName("Invocable")
 
@@ -52,25 +65,65 @@ internal val myChatMemberUpdateClass = MyChatMemberUpdate::class.asTypeName()
 internal val chatMemberUpdateClass = ChatMemberUpdate::class.asTypeName()
 internal val chatJoinRequestUpdateClass = ChatJoinRequestUpdate::class.asTypeName()
 
-internal fun List<KSValueArgument>.parseToCommandHandlerArgs() = Triple(
+internal val chainLinkClass = ChLink::class.asTypeName()
+
+internal fun List<KSValueArgument>.parseAsCommandHandler() = Triple(
     get(0).value.cast<List<String>>(),
     get(1).value.cast<KSAnnotation>().arguments.let { it.first().value.cast<Long>() to it.last().value.cast<Long>() },
     last().value.cast<List<KSType>>().map { UpdateType.valueOf(it.declaration.toString()) },
 )
 
-internal fun List<KSValueArgument>.parseToInputHandlerArgs() = Pair(
+internal fun List<KSValueArgument>.parseAsInputHandler() = Pair(
     get(0).value.cast<List<String>>(),
     get(1).value.cast<KSAnnotation>().arguments.let { it.first().value.cast<Long>() to it.last().value.cast<Long>() },
 )
 
-internal fun List<KSValueArgument>.parseToRegexHandlerArgs() = Pair(
+internal fun List<KSValueArgument>.parseAsRegexHandler() = Pair(
     get(0).value.cast<String>(),
     get(1).value.cast<KSAnnotation>().arguments.let { it.first().value.cast<Long>() to it.last().value.cast<Long>() },
 )
 
-internal fun List<KSValueArgument>.parseToUpdateHandlerArgs() = first().value.cast<List<KSType>>().map {
+internal fun List<KSValueArgument>.parseAsUpdateHandler() = first().value.cast<List<KSType>>().map {
     UpdateType.valueOf(it.declaration.toString())
 }
 
+internal fun List<KSValueArgument>.parseAsLinkOptions() = Pair(
+    get(0).value.cast<Boolean>(),
+    get(1).value.cast<KSAnnotation>().arguments.let { it.first().value.cast<Long>() to it.last().value.cast<Long>() },
+)
+
 @Suppress("UNCHECKED_CAST", "NOTHING_TO_INLINE")
 internal inline fun <R> Any?.cast(): R = this as R
+
+internal fun Pair<Long, Long>.toRateLimits() = RateLimits(first, second)
+
+internal fun <T : Annotation> Resolver.getAnnotatedFnSymbols(clazz: KClass<T>) =
+    getSymbolsWithAnnotation(clazz.qualifiedName!!).filterIsInstance<KSFunctionDeclaration>()
+
+internal fun <T : Annotation> Resolver.getAnnotatedClassSymbols(clazz: KClass<T>) =
+    getSymbolsWithAnnotation(clazz.qualifiedName!!).filterIsInstance<KSClassDeclaration>()
+
+internal fun FileBuilder.addMap(
+    name: String,
+    type: ParameterizedTypeName,
+    symbols: Sequence<KSFunctionDeclaration>,
+    tailBlock: CodeBlock? = null,
+    iterableAction: CodeBlock.Builder.(KSFunctionDeclaration) -> Unit,
+) = addProperty(
+    PropertySpec.builder(
+        name,
+        type,
+        KModifier.PRIVATE,
+    ).apply {
+        initializer(
+            CodeBlock.builder().apply {
+                add("mapOf(\n")
+                if (symbols.iterator().hasNext()) symbols.forEach {
+                    iterableAction(it)
+                }
+                tailBlock?.also { add(it) }
+                add(")\n")
+            }.build(),
+        )
+    }.build(),
+)
