+++ publishing.gradle.kts
@@ -13,7 +13,6 @@ configure<PublishingExtension> {
 
             pom {
                 name.set(artifactId)
-                description.set("Telegram Bot API wrapper, with handy Kotlin DSL.")
                 url.set("https://github.com/vendelieu/telegram-bot")
                 licenses {
                     license {
