+++ telegram-bot/build.gradle.kts
@@ -42,6 +42,7 @@ dependencies {
 
 group = "eu.vendeli"
 version = providers.gradleProperty("libVersion").getOrElse("dev")
+description = "Telegram Bot API wrapper, with handy Kotlin DSL."
 
 apply(from = "../publishing.gradle.kts")
