+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/annotations/InputChain.kt
@@ -1,0 +1,5 @@
+package eu.vendeli.tgbot.annotations
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.SOURCE)
+annotation class InputChain
