+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/implementations/UserDataMapImpl.kt
@@ -5,7 +5,7 @@ import eu.vendeli.tgbot.utils.asyncAction
 import kotlinx.coroutines.Deferred
 import java.util.concurrent.ConcurrentHashMap
 
-object UserDataMapImpl : UserData {
+class UserDataMapImpl : UserData {
     private val storage by lazy { ConcurrentHashMap<String, Any?>() }
     override fun set(telegramId: Long, key: String, value: Any?) {
         storage["$telegramId-$key"] = value
